Remove scope / does not build
