+++ sdk/trace/trace.go
@@ -19,7 +19,6 @@ import (
 	"math/rand"
 
 	"github.com/open-telemetry/opentelemetry-go/api/core"
-	"github.com/open-telemetry/opentelemetry-go/api/scope"
 	"github.com/open-telemetry/opentelemetry-go/api/tag"
 	"github.com/open-telemetry/opentelemetry-go/api/trace"
 	apitrace "github.com/open-telemetry/opentelemetry-go/api/trace"
